@@ -24,17 +24,17 @@
     }
   ],
   "require": {
-    "php": "^8.0",
-    "illuminate/support": "^9.21|^10.0",
-    "illuminate/config": "^10.21",
-    "illuminate/view": "^10.21",
-    "laravelcollective/html": "v6.4.1"
+    "php": ">=8.2",
+    "illuminate/support": "^11.0",
+    "illuminate/config": "^11.0",
+    "illuminate/view": "^11.0",
+    "spatie/laravel-html": "^3.9"
   },
   "require-dev": {
-    "phpunit/phpunit": "^10.1",
-    "mockery/mockery": "~1.5.1",
-    "orchestra/testbench": "^v8.10.0",
-    "friendsofphp/php-cs-fixer": "^3.21.1"
+    "phpunit/phpunit": "^10.5",
+    "mockery/mockery": "^1.6",
+    "orchestra/testbench": "^9.0",
+    "friendsofphp/php-cs-fixer": "^3.54"
   },
   "autoload": {
     "psr-4": {
@@ -48,7 +48,7 @@
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "6.0-dev"
+      "dev-master": "9.0-dev"
     },
     "laravel": {
       "providers": [
@@ -62,6 +62,7 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "scripts": {
-    "php-cs-fixer": "php-cs-fixer fix --config=./.php-cs-fixer.dist.php"
+    "php-cs-fixer": "php-cs-fixer fix --config=./.php-cs-fixer.dist.php",
+    "run-tests": "./vendor/bin/phpunit "
   }
 }
