@@ -4,6 +4,7 @@ namespace KyleMassacre\Menus\Tests;
 
 use KyleMassacre\Menus\Menu;
 use KyleMassacre\Menus\MenuBuilder;
+use PHPUnit\Framework\Attributes\Test;
 
 class MenuTest extends BaseTestCase
 {
@@ -17,8 +18,9 @@ class MenuTest extends BaseTestCase
         parent::setUp();
         $this->menu = app(Menu::class);
     }
-    /** @test */
-    public function it_generates_an_empty_menu()
+
+    #[Test]
+    public function it_generates_an_empty_menu(): void
     {
         $this->menu->create('test', function (MenuBuilder $menu) {
         });
@@ -34,8 +36,8 @@ TEXT;
         self::assertEquals($expected, $this->menu->get('test'));
     }
 
-    /** @test */
-    public function it_makes_is_an_alias_for_create()
+    #[Test]
+    public function it_makes_is_an_alias_for_create(): void
     {
         $this->menu->make('test', function (MenuBuilder $menu) {
         });
@@ -51,8 +53,8 @@ TEXT;
         self::assertEquals($expected, $this->menu->get('test'));
     }
 
-    /** @test */
-    public function it_render_is_an_alias_of_get()
+    #[Test]
+    public function it_render_is_an_alias_of_get(): void
     {
         $this->menu->make('test', function (MenuBuilder $menu) {
         });
@@ -68,8 +70,8 @@ TEXT;
         self::assertEquals($expected, $this->menu->render('test'));
     }
 
-    /** @test */
-    public function it_can_get_the_instance_of_a_menu()
+    #[Test]
+    public function it_can_get_the_instance_of_a_menu(): void
     {
         $this->menu->create('test', function (MenuBuilder $menu) {
         });
@@ -77,8 +79,8 @@ TEXT;
         $this->assertInstanceOf(MenuBuilder::class, $this->menu->instance('test'));
     }
 
-    /** @test */
-    public function it_can_modify_a_menu_instance()
+    #[Test]
+    public function it_can_modify_a_menu_instance(): void
     {
         $this->menu->create('test', function (MenuBuilder $menu) {
         });
@@ -90,8 +92,8 @@ TEXT;
         $this->assertCount(1, $this->menu->instance('test'));
     }
 
-    /** @test */
-    public function it_gets_a_partial_for_dropdown_styles()
+    #[Test]
+    public function it_gets_a_partial_for_dropdown_styles(): void
     {
         $this->menu->create('test', function (MenuBuilder $menu) {
         });
@@ -99,8 +101,8 @@ TEXT;
         $this->assertStringContainsString('.dropdown-submenu', $this->menu->style());
     }
 
-    /** @test */
-    public function it_can_get_all_menus()
+    #[Test]
+    public function it_can_get_all_menus(): void
     {
         $this->menu->create('main', function (MenuBuilder $menu) {
         });
@@ -110,8 +112,8 @@ TEXT;
         $this->assertCount(2, $this->menu->all());
     }
 
-    /** @test */
-    public function it_can_count_menus()
+    #[Test]
+    public function it_can_count_menus(): void
     {
         $this->menu->create('main', function (MenuBuilder $menu) {
         });
@@ -121,8 +123,8 @@ TEXT;
         $this->assertEquals(2, $this->menu->count());
     }
 
-    /** @test */
-    public function it_can_destroy_all_menus()
+    #[Test]
+    public function it_can_destroy_all_menus(): void
     {
         $this->menu->create('main', function (MenuBuilder $menu) {
         });
@@ -134,8 +136,8 @@ TEXT;
         $this->assertCount(0, $this->menu->all());
     }
 
-    /** @test */
-    public function it_still_generates_empty_menu_after_adding_dropdown()
+    #[Test]
+    public function it_still_generates_empty_menu_after_adding_dropdown(): void
     {
         $this->menu->create('test', function (MenuBuilder $menu) {
             $menu->dropdown('Test', function ($sub) {
@@ -156,8 +158,8 @@ TEXT;
         self::assertEquals($expected, $this->menu->get('test'));
     }
 
-    /** @test */
-    public function it_still_generates_empty_menu_after_adding_item()
+    #[Test]
+    public function it_still_generates_empty_menu_after_adding_item(): void
     {
         $this->menu->create('test', function (MenuBuilder $menu) {
             $menu->url('/', 'Test')
